% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heliclust.R
\name{heliclust_as_numeric}
\alias{heliclust_as_numeric}
\title{Internal helper: coerce all columns to numeric}
\usage{
heliclust_as_numeric(x)
}
\arguments{
\item{x}{Data frame or matrix.}
}
\description{
Internal helper: coerce all columns to numeric
}
\keyword{internal}
